+++ boards/posix/native_posix/CMakeLists.txt
@@ -14,6 +14,7 @@ zephyr_library_sources(
 	tracing.c
 	cmdline_common.c
 	cmdline.c
+	cpu_wait.c
 	hw_counter.c
 	)
