+++ boards/posix/nrf52_bsim/CMakeLists.txt
@@ -23,7 +23,7 @@ zephyr_compile_options(
 
 zephyr_library_sources(
 	irq_handler.c
-	k_busy_wait.c
+	cpu_wait.c
 	bstests_entry.c
 	argparse.c
 	main.c
