+++ drivers/timer/native_posix_timer.c
@@ -115,40 +115,6 @@ uint32_t z_clock_elapsed(void)
 }
 
 
-#if defined(CONFIG_ARCH_HAS_CUSTOM_BUSY_WAIT)
-/**
- * Replacement to the kernel k_busy_wait()
- * Will block this thread (and therefore the whole zephyr) during usec_to_wait
- *
- * Note that interrupts may be received in the meanwhile and that therefore this
- * thread may loose context
- *
- * This special arch_busy_wait() is necessary due to how the POSIX arch/SOC INF
- * models a CPU. Conceptually it could be thought as if the MCU was running
- * at an infinitely high clock, and therefore no simulated time passes while
- * executing instructions(*1).
- * Therefore to be able to busy wait this function does the equivalent of
- * programming a dedicated timer which will raise a non-maskable interrupt,
- * and halting the CPU.
- *
- * (*1) In reality simulated time is simply not advanced just due to the "MCU"
- * running. Meaning, the SW running on the MCU is assumed to take 0 time.
- */
-void arch_busy_wait(uint32_t usec_to_wait)
-{
-	uint64_t time_end = hwm_get_time() + usec_to_wait;
-
-	while (hwm_get_time() < time_end) {
-		/*
-		 * There may be wakes due to other interrupts including
-		 * other threads calling arch_busy_wait
-		 */
-		hwtimer_wake_in_time(time_end);
-		posix_halt_cpu();
-	}
-}
-#endif
-
 #if defined(CONFIG_SYSTEM_CLOCK_DISABLE)
 /**
  *
