+++ soc/posix/inf_clock/posix_soc.h
@@ -16,6 +16,7 @@ extern "C" {
 void posix_interrupt_raised(void);
 void posix_boot_cpu(void);
 int  posix_is_cpu_running(void);
+void posix_change_cpu_state_and_wait(bool halted);
 
 #ifdef __cplusplus
 }
